+++ src/test/resources/META-INF/services/org.testng.ITestNGListener
@@ -1,1 +1,1 @@
-#listeners.TestListener
+listeners.TestListener
