+++ src/test/resources/leapfrog.properties
@@ -1,0 +1,1 @@
+BASE_URL=https://store.leapfrog.com/en-us/apps/c?p=1&platforms=197&product_list_dir=asc&product_list_order=name
